Update Word Guessing Game
removed the function for target word and instead just defined a target word.  Added an else if statement to check if the user's input is longer than one character (one word guess) this way they can guess a whole word instead of one letter.

Still need to add a function to actually play the game and add in player turns.